requirements.txt
asgiref==3.7.2
Django==4.2.3
sqlparse==0.4.4
typing_extensions==4.7.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
django-adminlte-3==0.1.6
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-cors-headers==4.0.0
django-filter==23.2
django-js-asset==2.0.0
django-resized==1.0.2
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
docopt==0.6.2
drf-yasg==1.21.5
future==0.18.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pipreqs==0.4.13
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
serializers==0.2.4
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
yarg==0.1.9